requirements.txt
streamlit
rembg
Pillow
watchdog
onnxruntime
google-genai
python-dotenv
